Move solution-blocks-specific code from Artist into a "concern"
Which we then apply to Blockly.

Given that all blockly levels now have solution blocks, artist should no
longer be special. However, the blockly model is getting pretty
bloated, so this seemed like a good place to enforce some separation of
concerns.
